requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
awscli==1.32.108
botocore==1.34.108
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
couchbase==4.3.5
dataclasses-json==0.6.7
distro==1.9.0
docutils==0.16
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gritql==0.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-cli==0.0.35
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langserve==0.3.1
langsmith==0.3.6
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.61.1
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pyasn1==0.6.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.7.2
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
sse-starlette==1.8.2
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.18
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
